+++ pkg/alertmanager/validation/v1beta1/validation.go
@@ -173,27 +173,9 @@ func validateOpsGenieConfigs(configs []monitoringv1beta1.OpsGenieConfig) error {
 }
 
 func validateSlackConfigs(configs []monitoringv1beta1.SlackConfig) error {
-	for i, config := range configs {
+	for _, config := range configs {
 		if err := config.Validate(); err != nil {
 			return err
-		}
-
-		if err := validation.ValidateURLPtr((*string)(config.IconURL)); err != nil {
-			return fmt.Errorf("[%d]: iconURL: %w", i, err)
-		}
-
-		if err := validation.ValidateURLPtr((*string)(config.ImageURL)); err != nil {
-			return fmt.Errorf("[%d]: imageURL: %w", i, err)
-		}
-
-		if err := validation.ValidateURLPtr((*string)(config.ThumbURL)); err != nil {
-			return fmt.Errorf("[%d]: thumbURL: %w", i, err)
-		}
-
-		for j, sa := range config.Actions {
-			if err := validation.ValidateURLPtr((*string)(sa.URL)); err != nil {
-				return fmt.Errorf("[%d]: invalid 'action'[%d]: url: %w", i, j, err)
-			}
 		}
 
 		if err := config.HTTPConfig.Validate(); err != nil {